% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spectrumFunctions.R
\name{alignSpectra}
\alias{alignSpectra}
\title{Function to align a spectra to a reference spectrum, e.g. library spectrum}
\usage{
alignSpectra(x, y, mzTol = 0.005, mzTolType = "abs", treshold = 0.01,
  returnSpectra = FALSE)
}
\description{
Function to align a spectra to a reference spectrum, e.g. library spectrum
}
