% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_SpectrumFunctions.R
\name{bin_Spectra}
\alias{bin_Spectra}
\title{Function to bin spectra on common bins}
\usage{
bin_Spectra(x, y, binSize = 1L, breaks = seq(floor(min(c(mz(x),
  mz(y)))), ceiling(max(c(mz(x), mz(y)))), by = binSize),
  returnSpectra = FALSE)
}
\description{
Function to bin spectra on common bins
}
