% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{calcAdductFormula}
\alias{calcAdductFormula}
\title{This function generates the formula of a pseudomolecular ion from a chemical formula and the type of adduct selected.}
\usage{
calcAdductFormula(molecularFormula, adduct)
}
\value{
Returns a string with the ion formula
}
\description{
This function generates the formula of a pseudomolecular ion from a chemical formula and the type of adduct selected.
}
\examples{
ionFormula <- calcAdductFormula("C6H12O6", "M+H")
}
