% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{calculateExactMass}
\alias{calculateExactMass}
\title{This function calculates the exact mass from a given sum formula.}
\usage{
calculateExactMass(chemFormula)
}
\arguments{
\item{chemFormula}{Single string with chemical formula}
}
\value{
calculated exact mass
}
\description{
This function calculates the exact mass from a given sum formula.
}
\examples{
calculateExactMass("C6H12O6")
}
