% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{calculateKendrickMass}
\alias{calculateKendrickMass}
\title{Calculate the Kendrick Mass}
\usage{
calculateKendrickMass(mz)
}
\arguments{
\item{mz}{Mass for which the KM shall be calculated}
}
\description{
This function calculates the Kendrick mass for a give mass.
}
\seealso{
\code{\link{calculateRefKendrickMassDefect}}

\code{\link{calculateKendrickMassDefect}}

\code{\link{calculateNominalKendrickMass}}

\code{\link{checkRmkd}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
