% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{calculateKendrickMassDefect}
\alias{calculateKendrickMassDefect}
\title{Calculate Kendrick Mass Defect}
\usage{
calculateKendrickMassDefect(mz)
}
\arguments{
\item{mz}{Mass for which the KMD shall be calculated}
}
\description{
This function calculates the Kendrick mass defect (KMD) for a given mass.
}
\seealso{
\code{\link{calculateRefKendrickMassDefect}}

\code{\link{checkRmkd}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
