% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{calculateNominalKendrickMass}
\alias{calculateNominalKendrickMass}
\title{Calculate nominal Kendrick Mass}
\usage{
calculateNominalKendrickMass(mz)
}
\arguments{
\item{mz}{Mass for which the nominal KM shall be calculated}
}
\description{
This function calculates the nominal Kendrick Mass
}
