% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{calculateRefKendrickMassDefect}
\alias{calculateRefKendrickMassDefect}
\title{Calculate referenced Kendrick Mass Defect}
\usage{
calculateRefKendrickMassDefect(mz, refkmd)
}
\arguments{
\item{mz}{Mass for which the RKMD shall be calculated}

\item{refkmd}{Reference Kendrick Mass Defect}
}
\description{
This function calculates a referenced Kendrick Mass Defect (RKMD) for a given mass and a given reference KMD.
}
\examples{
xxx

}
\seealso{
\code{\link{calculateKendrickMassDefect}}

\code{\link{checkRmkd}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
