% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{checkRmkd}
\alias{checkRmkd}
\title{Check if RMKD is negative integer}
\usage{
checkRmkd(rmkd, error = 0.15)
}
\arguments{
\item{rmkd}{calculated RMKD}

\item{error}{allowed error margin for RMKD}
}
\description{
This function checks if the RKMD is a negative integer
}
\seealso{
\code{\link{calculateKendrickMassDefect}}

\code{\link{calculateRefKendrickMassDefect}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
