% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_Utilities.R
\name{clean_spectrum}
\alias{clean_spectrum}
\title{This function is used to clean MS2 spectra in different ways. First, if the MS2 contains "fragment" m/z values above the precursor m/z (option "removeAbovePrecursor"). Second, isotopic peaks in the MS2 spectra can be removed ("isotopeFilitering"). Lastly, peaks below a certain treshold are removed ("tresholdFiltering").}
\usage{
clean_spectrum(
  x,
  cleanFunction = "removeAbovePrecursor",
  treshold,
  precursorDistance = 5,
  mzTol = 0.005,
  mzTolType = "abs"
)
}
\arguments{
\item{x}{A Spectrum2 object to work on}

\item{cleanFunction}{A string either "removeAbovePrecursor", "isotopeFiltering" or "tresholdFiltering"}

\item{precursorDistance}{A double indicating from which point above which "fragments" should be removed}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}
}
\description{
This function is used to clean MS2 spectra in different ways. First, if the MS2 contains "fragment" m/z values above the precursor m/z (option "removeAbovePrecursor"). Second, isotopic peaks in the MS2 spectra can be removed ("isotopeFilitering"). Lastly, peaks below a certain treshold are removed ("tresholdFiltering").
}
