% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpecSimilarty.R
\name{common_peaks}
\alias{common_peaks}
\title{Calculate common peaks}
\usage{
common_peaks(x, y, align = FALSE, mzTol = 0.005, mzTolType = "abs",
  binwidth = 1L, ...)
}
\description{
Calculate common peaks
}
