% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpecSimilarty.R
\name{commonPeaks}
\alias{commonPeaks}
\title{Calculate common peaks}
\usage{
commonPeaks(x, y, align = FALSE, mzTol = 0.005, mzTolType = "abs",
  binwidth = 1L, ...)
}
\description{
Calculate common peaks
}
