% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_Utilities.R
\name{containsFragmentDifference}
\alias{containsFragmentDifference}
\title{This function allows to search for a specific mass difference between two fragments}
\usage{
containsFragmentDifference(ms2spectrum, fragmentMassDifference,
  mzTol = 0.005, mzTolType = "abs")
}
\arguments{
\item{ms2spectrum}{A Spectrum2 in which the production ion shall be searched}

\item{fragmentMassDifference}{Mass of neutral loss that shal be searched}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}
}
\description{
This function allows to search for a specific mass difference between two fragments
}
