% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_Utilities.R
\name{containsNeutralLossIon}
\alias{containsNeutralLossIon}
\title{This function allows to search for a fragment m/z in MS2 spectra that match to a specified neutral from the precusor}
\usage{
containsNeutralLossIon(ms2spectrum, neutralLossMass, mzTol = 0.005,
  mzTolType = "abs")
}
\arguments{
\item{ms2spectrum}{A Spectrum2 in which the production ion shall be searched}

\item{neutralLossMass}{Mass of neutral loss that shal be searched}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}
}
\description{
This function allows to search for a fragment m/z in MS2 spectra that match to a specified neutral from the precusor
}
