% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_Utilities.R
\name{containsProductIon}
\alias{containsProductIon}
\title{This function allows to search for a specific product ion in MS2 spectra}
\usage{
containsProductIon(ms2spectrum, productIonMz, mzTol = 0.005,
  mzTolType = "abs", multiplePi = "any")
}
\arguments{
\item{ms2spectrum}{A Spectrum2 in which the production ion shall be searched}

\item{productIonMz}{m/z value of product ion that shall be search for}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}
}
\description{
This function allows to search for a specific product ion in MS2 spectra
}
