% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpectrumDbSearch.R
\name{create_results_set}
\alias{create_results_set}
\title{Function that calculates scores for a query and a list of result spectra}
\usage{
create_results_set(
  query_spectrum,
  result_spectra,
  align = TRUE,
  mzTol = 0.005,
  mzTolType = "abs",
  prefix = ""
)
}
\arguments{
\item{querySpectrum}{A Spectrum2 object}

\item{queryResults}{A Spectra object containing the results from a DB search}
}
\description{
Function that calculates scores for a query and a list of result spectra
}
