% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_AdductDbCreate.R
\name{createDb}
\alias{createDb}
\title{Creation of SQLiteDB}
\usage{
createDb(compoundList, dbName)
}
\arguments{
\item{compoundList}{List of compounds that shall be added to DB}

\item{dbName}{A name for the database file}
}
\description{
masstrixR uses SQLite DBs as backend for the annotation process. In order to perform annotation of m/z values with putative metabolites a DB is required.
This function creates a SQLite database using a validated compound list containing all required information. Such a list can be created using the \code{prepareCompoundList()} function.
}
\examples{
xxx

}
\seealso{
\code{\link{validateCompoundList}}

\code{\link{prepareCompoundList}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
