% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpectrumDbCreate.R
\name{createMs2Db}
\alias{createMs2Db}
\title{This function creates a MS2 library that can be used for MS2 database searches.}
\usage{
createMs2Db(librarySpectra, ms2dbFileName)
}
\arguments{
\item{librarySpectra}{A Spectra object with fitting metadata columns (e.g. generated by the <code>readMassBankFile</code> function)}

\item{ms2dbFileName}{File name that shall be used for the database.}
}
\description{
This function creates a MS2 library that can be used for MS2 database searches.
}
