% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_spectrumFunctions.R
\name{.bin_values}
\alias{.bin_values}
\title{binngin function}
\usage{
.bin_values(x, toBin, binSize = 1, breaks = seq(floor(min(toBin)),
  ceiling(max(toBin)), by = binSize), fun = max)
}
\description{
binngin function
}
