% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpecSimilarty.R
\name{dotproduct}
\alias{dotproduct}
\title{Helper function to calculate dot product between two intensity vectors on same m/z scale (binned or aligned)}
\usage{
dotproduct(x, y)
}
\description{
Helper function to calculate dot product between two intensity vectors on same m/z scale (binned or aligned)
}
