% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_Utilities.R
\name{filterIsotopePeaks}
\alias{filterIsotopePeaks}
\title{This function uses CAMERA to identify isotopes in a mz intensity pair.}
\usage{
filterIsotopePeaks(mz, int, mzTol = 0.005, mzTolType = "abs")
}
\arguments{
\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}
}
\description{
This function uses CAMERA to identify isotopes in a mz intensity pair.
}
