% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_specSimilarty.R
\name{forward_dotproduct}
\alias{forward_dotproduct}
\title{Calculate forward Dotproduct}
\usage{
forward_dotproduct(
  x,
  y,
  align = FALSE,
  mzTol = 0.005,
  mzTolType = "abs",
  binwidth = 1L,
  ...
)
}
\description{
Calculate forward Dotproduct
}
