% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_formula.R
\name{generateChemFormula}
\alias{generateChemFormula}
\title{This function generates a chemical formula from a named vector of elemental counts}
\usage{
generateChemFormula(parsedChemFormula)
}
\value{
Single string with chemical formula
}
\description{
This function generates a chemical formula from a named vector of elemental counts
}
\examples{
parseChemFormula("C6H12O6")
}
