% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_AdductCalc.R
\name{getAdductCalc}
\alias{getAdductCalc}
\title{This function returns a list with adducts and the required multiplicative and additive part to calculate ion m/z from exact masses. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator}
\usage{
getAdductCalc()
}
\value{
Returns a named list with all mulitpliers and additive parts for positive and negative mode
}
\description{
This function returns a list with adducts and the required multiplicative and additive part to calculate ion m/z from exact masses. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator
}
\examples{
adductCalc <- getAdductCalc()
}
