% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_AdductCalc.R
\name{getAllNegModeAdducts}
\alias{getAllNegModeAdducts}
\title{This function returns a list with the names of the adducts used to call the fitting formula. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator}
\usage{
getAllNegModeAdducts()
}
\value{
Returns a list with all adduct names for the negative ionization mode
}
\description{
This function returns a list with the names of the adducts used to call the fitting formula. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator
}
\examples{
getAllNegModeAdducts
}
