% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adductCalc.R
\name{getAllPosModeAdducts}
\alias{getAllPosModeAdducts}
\title{This function returns a list with the names of the adducts used to call the fitting formula. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator}
\usage{
getAllPosModeAdducts()
}
\value{
Returns a list with all adduct names for the positive ionization mode
}
\description{
This function returns a list with the names of the adducts used to call the fitting formula. Adduts are taken from http://fiehnlab.ucdavis.edu/staff/kind/metabolomics/ms-adduct-calculator
}
\examples{
getAllPosModeAdducts()
}
