% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_ApiCts.R
\name{getExternalDbIds}
\alias{getExternalDbIds}
\title{This function uses the CTS Webservice from the Fiehnlab to convert IDs}
\usage{
getExternalDbIds(from, to, queryString)
}
\arguments{
\item{from}{format of original identifier (kegg, hmdb, inchikey)}

\item{to}{format of identifier to convert to}

\item{queryString}{query that shall be converted}
}
\description{
This function uses the CTS Webservice from the Fiehnlab to convert IDs
}
