% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{getIsotopeAddList}
\alias{getIsotopeAddList}
\title{function to get the differences between monoisotope and isotope}
\usage{
getIsotopeAddList()
}
\description{
function to get the differences between monoisotope and isotope
}
