% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_adductCalc.R
\name{getNeutralLossCalc}
\alias{getNeutralLossCalc}
\title{This function returns a list with typical neutral losses that can be observed as in-source fragments}
\usage{
getNeutralLossCalc()
}
\value{
Returns a named list with all mulitpliers and additive parts for positive and negative mode
}
\description{
This function returns a list with typical neutral losses that can be observed as in-source fragments
}
\examples{
neutralLossCalc <- getNeutralLossCalc()
}
