% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_IsotopePattern.R
\name{isoPatternSimilarity}
\alias{isoPatternSimilarity}
\title{Calculate similarity of isotope pattern}
\usage{
isoPatternSimilarity(x, y, mzTol = 0.005, ...)
}
\description{
based on S-ratio from MS-Dial
}
