% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{matchMassDiff}
\alias{matchMassDiff}
\title{Checking for mass diferences}
\usage{
matchMassDiff(mz1, mz2, mzDiff, mzTol = 0.005, mzTolType = "abs")
}
\arguments{
\item{mz1}{First m/z value for which a mass difference shall be checked}

\item{mz2}{Second m/z value for which a mass difference shall be checked}

\item{mzDiff}{Mass difference to check for}

\item{mzTol}{m/z error, numeric value}

\item{mzTolType}{type of error used, absolute (abs) or relative (ppm)}
}
\description{
This functions is used to check of two measured masses are different by a give mass difference. This can be used to check if a neutral loss is found between two masses or if masses are connected by certain metabolic transformations, e.g. methylation, acetylation etc.
}
\examples{
xxx

}
\seealso{
\code{\link{matchMz}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
