% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_Utilities.R
\name{matchMz}
\alias{matchMz}
\title{This function is used to match two masses from the positive and negative ionization mode to check it they are potentially derived from the same metabolite. Based on the given adducts, all combinations are tested from positive to negative ion mode and the other way. If for both cases the mass error is below the given mass error it is assumed that the masses are derived from the same metabolite.}
\usage{
matchMz(posMz, negMz, posAdducts, negAdducts, mzTol = 0.005,
  mzTolType = "abs")
}
\arguments{
\item{posMz}{m/z value of positive ionization mode}

\item{negMz}{m/z value of negative ionization mode}

\item{posAdducts}{vector with adducts that are used for the positive ionization mode}

\item{negAdducts}{vector with adducts that are used for the negative ionization mode}

\item{mzTol}{m/z error, numeric value}

\item{mzTolType}{type of error used, absolute (abs) or relative (ppm)}
}
\description{
This function is used to match two masses from the positive and negative ionization mode to check it they are potentially derived from the same metabolite. Based on the given adducts, all combinations are tested from positive to negative ion mode and the other way. If for both cases the mass error is below the given mass error it is assumed that the masses are derived from the same metabolite.
}
\examples{
xxx

}
\seealso{
\code{\link{matchMassDiff}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
