% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_mzSearch.R
\name{mzLookUp}
\alias{mzLookUp}
\title{This function annotates masses in a peak list with possible metabolites from a selected DB. Used adducts are the adducts covered by the respective database}
\usage{
mzLookUp(peakList, compoundList, mzTol = 0.005, mzTolType = "abs",
  rt = FALSE, rtTol = 0.5, rtTolType = "abs", ccs = FALSE,
  ccsTol = 1, ccsTolType = "rel")
}
\arguments{
\item{peakList}{Data frame containing the peaks of interest, minimum column is called m.z ($m.z)}

\item{mzTol}{Maximum allowed tolerance in Da (ppm will come soon)}

\item{dbFileName}{A .sqlite file containing the DB of interest}

\item{mode}{Defines if the database shall be used from the disk ("onDisk") or in memory ("inMemory"). The second option enhances performances with very large peaks lists and DBs.}
}
\value{
Returns the filename of the generated database
}
\description{
This function annotates masses in a peak list with possible metabolites from a selected DB. Used adducts are the adducts covered by the respective database
}
\examples{
mzLookup()

}
