% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_mzSearch.R
\name{mzSearch}
\alias{mzSearch}
\title{Performing m/z search}
\usage{
mzSearch(peakList, dbFileName, mode = "onDisk", mzTol = 0.005,
  mzTolType = "abs", rt = FALSE, rtTol = 0.5, rtTolType = "abs",
  ccs = FALSE, ccsTol = 1, ccsTolType = "rel")
}
\arguments{
\item{peakList}{Data frame containing the peaks of interest, minimum column is called m.z ($m.z)}

\item{dbFileName}{A .sqlite file containing the DB of interest}

\item{mode}{Defines if the database shall be used from the disk ("onDisk") or in memory ("inMemory"). The second option enhances performances with very large peaks lists and DBs.}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}

\item{rt}{Boolean value indicating of RT filtering shall be performed}
}
\value{
Returns the filename of the generated database
}
\description{
This function performs annotation of m/z values with putative metabolites using a previously created SQLite DB. Measured m/z values are compared against theoretical values within a defined error range. This error can be either absolute (in Da) or relative (in ppm).
If the database contains RT or CCS values results can be a additionally filtered.
}
\examples{
mzSearch()
}
