% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_isotopePattern.R
\name{predictIsoPattern}
\alias{predictIsoPattern}
\title{Prediction of isotope pattern}
\usage{
predictIsoPattern(ionFormula, charge, plotit = FALSE, treshold = 0.01,
  resolution = 50000)
}
\arguments{
\item{ionFormula}{The formula of the ion for which the isotopic pattern shall be calculated, e.g. C6H12O6Na for the M+Na adduct of glucose}

\item{charge}{Charge of the ion.}

\item{plotit}{Boolean value if the isotope pattern shall be plotted}

\item{treshold}{Intensity treshold for lowest isotope}

\item{resolution}{resolution of the MS}
}
\value{
Returns a MSnbase Spectrum 1 object containing the calculated isotope pattern
}
\description{
This function predicts the isotope pattern for a given ion formula and adduct definition.
}
\examples{
predictIsoPattern()
}
