% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_createAdductDb.R
\name{prepareCompoundList}
\alias{prepareCompoundList}
\title{Preparation of compound list}
\usage{
prepareCompoundList(compoundList, adductList = NA, rt = FALSE,
  ccs = FALSE, extId = FALSE)
}
\arguments{
\item{compoundList}{List of compounds that shall be added to DB}

\item{adductList}{Vector with adducts that shall be covered in the DB.}

\item{rt}{Boolean value indicating if compound list contains RT data}

\item{ccs}{Boolean value indicating if compound list contains CCS data}
}
\value{
Returns a data frame suitable for upload to a SQLite DB.
}
\description{
A compound list that can be used with masstrixR can be generated on the fly. Minimum input is a data frame with the following columns: metabolite id ($id), SMILES ($smiles), InChI ($inchi), InChIKey ($inchikey), formula ($formula) metabolite name ($name) and an exact mass ($exactmass).
Furthermore, the adducts that shall be covered in the DB have to be defined. This can be either done by using an list of adducts or supplying a adduct definition for each metabolite with an additiona adduct column ($adducts). If it is intended to perform retetion time and collisional cross section matching columns containing this information are required ($rt and $ccs).
In case of CCS matching, individual adduct rules are required since each adduct has a different CCS value. Examples for each input are found in the examples in the vignettes.
}
\examples{
prepareCompoundList(compoundList, adducts = c("M+H", "M+Na"))

}
\seealso{
\code{\link{validateCompoundList}}

\code{\link{createDb}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
