% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_AdductDbCreate.R
\name{prepareIsoCompoundList}
\alias{prepareIsoCompoundList}
\title{iso list}
\usage{
prepareIsoCompoundList(compoundList, isoLabel = "full", labeledElement,
  noOfLabel = NULL)
}
\description{
iso list
}
