% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_parseGda.R
\name{readGdaFile}
\alias{readGdaFile}
\title{This funtion parses a Genedata .gda file and returns three data frames containing the data and row and column annotations}
\usage{
readGdaFile(pathToFile)
}
\arguments{
\item{pathTofile}{Path to .gda file that shall be loaded}
}
\value{
Returns a list with the three different dataframes, index 1 is the data, index 2 column annotation, index 3 row annotations
}
\description{
This funtion parses a Genedata .gda file and returns three data frames containing the data and row and column annotations
}
\examples{
xxx
}
