% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_IsotopePattern.R
\name{reconstructIsoPattern}
\alias{reconstructIsoPattern}
\title{Reconstruct isotope pattern from gda files}
\usage{
reconstructIsoPattern(peaks, sampleNames)
}
\description{
Reconstruct isotope pattern from gda files
}
