% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_specSimilarty.R
\name{reverse_dotproduct}
\alias{reverse_dotproduct}
\title{Calculate reverse Dotproduct}
\usage{
reverse_dotproduct(x, y, align = FALSE, mzTol = 0.005,
  mzTolType = "abs", binwidth = 1L, ...)
}
\description{
Calculate reverse Dotproduct
}
