% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_specSimilarty.R
\name{reverseDotProduct}
\alias{reverseDotProduct}
\title{Calculate reverse Dotproduct}
\usage{
reverseDotProduct(x, y, align = FALSE, mzTol = 0.005,
  treshold = 0.01, ...)
}
\description{
Calculate reverse Dotproduct
}
