% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_IsotopePattern.R
\name{scoreIsoPattern}
\alias{scoreIsoPattern}
\title{Function to score isotope pattern}
\usage{
scoreIsoPattern(filteredAnnotationResult, measuredIsotopePattern,
  returnPredicted = FALSE)
}
\description{
Function to score isotope pattern
}
