% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_MoNAAPI.R
\name{searchByExactMassOnline}
\alias{searchByExactMassOnline}
\title{This function performs a exact mass search in MoNA using an httr GET request and returns results a list. Data in this list can be access via the functions in parsingFunctions.R.}
\usage{
searchByExactMassOnline(exactMass, error = 0.01, errorType = "abs")
}
\arguments{
\item{exactMass}{exact mass to search for}

\item{error}{numeric value for maximum error}

\item{errorType}{type of error for search, either abs (absolute error in Da) or ppm (ppm error)}
}
\value{
Returns a List of lists with all information available (use parsingFunctions.R for simple data access)
}
\description{
This function performs a exact mass search in MoNA using an httr GET request and returns results a list. Data in this list can be access via the functions in parsingFunctions.R.
}
