% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_MoNAAPI.R
\name{searchByFormulaOnline}
\alias{searchByFormulaOnline}
\title{This function performs a formula search in MoNA using an httr GET request and returns results a list. Data in this list can be access via the functions in parsingFunctions.R.}
\usage{
searchByFormulaOnline(formula)
}
\arguments{
\item{formula}{Formula to search for}
}
\value{
Returns a List of lists with all information available (use parsingFunctions.R for simple data access)
}
\description{
This function performs a formula search in MoNA using an httr GET request and returns results a list. Data in this list can be access via the functions in parsingFunctions.R.
}
