% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_ApiMoNA.R
\name{searchByPrecursorOnline}
\alias{searchByPrecursorOnline}
\title{This function reconstructs a Ms2 spectrum for all fragments with a correlation value bigger than a set cut-off}
\usage{
searchByPrecursorOnline(precursor, error = 0.01, errorType = "abs",
  precursorType = "[M+H]+")
}
\arguments{
\item{error}{numeric value for maximum error}

\item{errorType}{type of error for search, either abs (absolute error in Da) or ppm (ppm error)}

\item{precursorType}{Indicates possible adduct ions to search for (currently supported: [M+H]+, [M-H]-)}

\item{exactMass}{exact mass to search for}
}
\value{
Returns a List of lists with all information available (use parsingFunctions.R for simple data access)
}
\description{
This function reconstructs a Ms2 spectrum for all fragments with a correlation value bigger than a set cut-off
}
