% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{standardizeChemFormula}
\alias{standardizeChemFormula}
\title{This function standardizes a supplied chemical formula according to the Hill notation system.}
\usage{
standardizeChemFormula(chemFormula)
}
\arguments{
\item{chemFormula}{Single string with chemical formula}
}
\value{
Single string with standardized chemical formua
}
\description{
This function standardizes a supplied chemical formula according to the Hill notation system.
}
\examples{
standardizeChemFormula("O6C6H12")
}
