% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_createAdductDb.R
\name{validateCompoundList}
\alias{validateCompoundList}
\title{Validation of compound list}
\usage{
validateCompoundList(compoundList, rt = FALSE, ccs = FALSE,
  extId = FALSE)
}
\arguments{
\item{compoundList}{List of compounds that shall be added to DB}
}
\value{
Returns the filename of the generated database
}
\description{
Compound list that are used to generate SQLite DBs require a speficic format prior to generation of the actual DB.
}
\details{
This function validates if all required columns are in place and of the right format. Specific columns are added, if required (e.g. KEGG Ids)
}
\examples{
xxx

}
\seealso{
\code{\link{createDb}}

\code{\link{prepareCompoundList}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
