% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_SpectrumFunctions.R
\name{bin_Spectra}
\alias{bin_Spectra}
\title{Function to bin spectra on common bins}
\usage{
bin_Spectra(object1, object2, binSize = 1L,
  breaks = seq(floor(min(c(mz(object1), mz(object2)))),
  ceiling(max(c(mz(object1), mz(object2)))), by = binSize))
}
\description{
Function to bin spectra on common bins
}
