% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{containsFormula}
\alias{containsFormula}
\title{This function checks if one sum formula is contained in another.}
\usage{
containsFormula(targetFormula, queryFormula)
}
\arguments{
\item{targetFormula}{Single string with chemical formula}

\item{queryFormula}{Single string with chemical formula that should be contained in the targetFormula}
}
\value{
TRUE or FALSE
}
\description{
This function checks if one sum formula is contained in another.
}
\examples{
ionFormula <- containsFormula("C6H12O6", "H2O")
}
