% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpecSimilarty.R
\name{forwardDotProduct}
\alias{forwardDotProduct}
\title{Calculate forward Dotproduct}
\usage{
forwardDotProduct(x, y, align = FALSE, mzTol = 0.005,
  mzTolType = "abs", binwidth = 1L, ...)
}
\description{
Calculate forward Dotproduct
}
