% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{generateChemFormula}
\alias{generateChemFormula}
\title{This function generates a chemical formula from a named vector of elemental counts}
\usage{
generateChemFormula(parsedChemFormula)
}
\arguments{
\item{parsedChemFormula}{Named list containing the elements as names and their abudance as values}
}
\value{
Single string with chemical formula
}
\description{
This function generates a chemical formula from a named vector of elemental counts
}
\examples{
formula <- list("C" = 6,
                "H" = 12,
                "O" = 6)
generateChemFormula(formula)
}
