% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS1_AdductDbSearch.R
\name{mzLookUp}
\alias{mzLookUp}
\title{Performing m/z search (fast coding)}
\usage{
mzLookUp(peakList, compoundList, mzTol = 0.005, mzTolType = "abs",
  rt = FALSE, rtTol = 0.5, rtTolType = "abs", ccs = FALSE,
  ccsTol = 1, ccsTolType = "rel")
}
\arguments{
\item{peakList}{Data frame containing the peaks of interest, minimum column is called mz ($mz) or m.z ($m.z) or mzmed ($mzmed)}

\item{compoundList}{A data frame with a validated compound list see \code{\link{validateCompoundList}}}

\item{mzTol}{Maximum allowed tolerance in Da or ppm}

\item{mzTolType}{Defines the error type for m/z search, "abs" is used for absolute mass error, "ppm" for relative error}

\item{rt}{Boolean value indicating of RT filtering shall be performed, if this option is true all entries in the DB need a RT value}

\item{rtTol}{Maximum allowed tolerance for RT search in the unit of the DB and peaklist or in relative error (%). Please note: Unit of peakList and DB have to match, e.g. both in seconds, minutes or RI units}

\item{rtTolType}{Defines the error type for RT search, "abs" is used for absolute RT error, "rel" for relative error}

\item{ccs}{Boolean value indicating of CCS filtering shall be performed, if this option is true all entries in the DB need a CCS value}

\item{ccsTol}{Maximum allowed tolerance for CCS search}

\item{rtTolType}{Defines the error type for CCS search, "abs" is used for absolute CCS error, "rel" for relative error}
}
\value{
Returns a data frame with results of the annotation. If for one single m/z value multiple annotations are found, each annotation is represented as individual row.
}
\description{
This function performs annotation of m/z values with putative metabolites. In comparison to the \code{mzSearch} function no SQLite DB is required. A SQLite DB is created in memory during the annotation process and deleted afterwards. Instead of the path of the a .sqlite file a data frame containing a validated compound list is required. Measured m/z values are compared against theoretical values within a defined error range. This error can be either absolute (in Da) or relative (in ppm).
If the database contains RT or CCS values results can be a additionally filtered.
}
\examples{
mzLookUp()

}
\seealso{
\code{\link{mzSearch}}
}
\author{
Michael Witting, \email{michael.witting@helmholtz-muenchen.de}
}
