% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils_Formula.R
\name{parseChemFormula}
\alias{parseChemFormula}
\title{This function parses a chemical formula into a named vector}
\usage{
parseChemFormula(chemFormula)
}
\arguments{
\item{chemFormula}{Single string with chemical formula}
}
\value{
Named vector with all elements
}
\description{
This function parses a chemical formula into a named vector
}
\examples{
parseChemFormula("C6H12O6")
}
