% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_LibraryReader.R
\name{readMassBankFile}
\alias{readMassBankFile}
\title{Function to read a single MassBank record
Inspeired from MassBank parser from RMassBank
https://github.com/sneumann/RMassBank/blob/master/R/parseMassBank.R}
\usage{
readMassBankFile(pathToMBFile)
}
\description{
Function to read a single MassBank record
Inspeired from MassBank parser from RMassBank
https://github.com/sneumann/RMassBank/blob/master/R/parseMassBank.R
}
