% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MS2_SpecSimilarty.R
\name{reverseDotProduct}
\alias{reverseDotProduct}
\title{Calculate reverse Dotproduct}
\usage{
reverseDotProduct(x, y, align = FALSE, mzTol = 0.005,
  mzTolType = "abs", binwidth = 1L, ...)
}
\description{
Calculate reverse Dotproduct
}
